fix efi boot order
When adding efi boot entries with `efibootmgr` the boot order is lost. In particular the current boot (pxe boot) is no longer the first boot item, so next time a user will want to deploy on this host, she will be forced to re-configure the boot order in the efi boot menu.
This fix move the current boot to the front of the list.